Fixed bug with syncing, implemented mobile network setting #105 #112

+++ ExperienceSamplingApp/app/src/main/java/com/example/madiskar/experiencesamplingapp/NetworkChangeReceiver.java
@@ -8,8 +8,10 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.ConnectivityManager;
+import android.net.Network;
 import android.net.NetworkInfo;
 import android.os.Handler;
+import android.preference.PreferenceManager;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -25,70 +27,84 @@ public class NetworkChangeReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(final Context context, Intent intent) {
-        Log.i("NETWORK STATE CHANGED", "CHECK IF CONNECTION IS AVAILABLE");
+        Log.i("NETWORK STATE CHANGED", "CHECK PARAMETERS");
 
         mContext = context;
         mydb = DBHandler.getInstance(context);
+
+        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
+        boolean mobileSyncAllowed = settings.getBoolean("pref_sync", false);
+        Log.i("mobile sync", String.valueOf(mobileSyncAllowed));
 
         SharedPreferences spref = context.getApplicationContext().getSharedPreferences("com.example.madiskar.ExperienceSampler", Context.MODE_PRIVATE);
         token = spref.getString("token", "none");
 
         ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
         NetworkInfo networkInfo = cm.getActiveNetworkInfo();
-
-        if (networkInfo != null && networkInfo.isConnected()) {
-            Log.i("NETWORK STATE CHANGED", "TRY TO SYNC DATA");
-
-            SyncResultDataTask syncResultDataTask = new SyncResultDataTask(true, mydb, token, new RunnableResponse() {
-                @Override
-                public void processFinish(String output) {
-                    Log.i("UPLOADED DATA:", output);
-                    Log.i("STARTING SYNC:", "Study info");
+        int networkType = -1;
+        try {
+            networkType = networkInfo.getType();
+        } catch (NullPointerException e) {
+            // do nothing
+        }
 
 
-                    SyncStudyDataTask syncStudyDataTask = new SyncStudyDataTask(token, mydb, true, new StudyDataSyncResponse() {
-                        @Override
-                        public void processFinish(String output, ArrayList<Study> newStudies, ArrayList<Study> allStudies, ArrayList<Study> updatedStudies, ArrayList<Study> oldStudies, ArrayList<Study> cancelledStudies) {
-                            if(output.equals("invalid_token")) {
-                                Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.auth_sync_fail));
-                            } else if(output.equals("nothing")) {
-                                Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.fetch_sync_fail));
-                            } else if(!output.equals("dberror")){
-                                for (int i=0; i<updatedStudies.size(); i++) {
-                                    Log.i("STUDIES MODIFIED", "notification data changed for " + updatedStudies.size() + " studies");
-                                    cancelStudy(oldStudies.get(i));
-                                    setUpNewStudy(updatedStudies.get(i));
-                                }
-                                for (Study s : cancelledStudies) {
-                                    Log.i("STUDIES CANCELLED", "removed " + cancelledStudies.size() + " studies");
-                                    for(Study ks : allStudies) {
-                                        if(ks.getId() == s.getId()) {
-                                            allStudies.remove(ks);
-                                            break;
-                                        }
+        if (networkInfo != null && networkInfo.isConnected() && !token.equals("none")) {
+            if(((networkType == ConnectivityManager.TYPE_MOBILE || networkType == ConnectivityManager.TYPE_MOBILE_DUN) && !mobileSyncAllowed) || networkType == ConnectivityManager.TYPE_DUMMY) {
+                Log.i("NETWORK STATE CHANGED", "DON'T SYNC");
+            } else {
+                Log.i("NETWORK STATE CHANGED", "TRY TO SYNC DATA");
+
+                SyncResultDataTask syncResultDataTask = new SyncResultDataTask(true, mydb, token, new RunnableResponse() {
+                    @Override
+                    public void processFinish(String output) {
+                        Log.i("UPLOADED DATA:", output);
+                        Log.i("STARTING SYNC:", "Study info");
+
+
+                        SyncStudyDataTask syncStudyDataTask = new SyncStudyDataTask(token, mydb, true, new StudyDataSyncResponse() {
+                            @Override
+                            public void processFinish(String output, ArrayList<Study> newStudies, ArrayList<Study> allStudies, ArrayList<Study> updatedStudies, ArrayList<Study> oldStudies, ArrayList<Study> cancelledStudies) {
+                                if (output.equals("invalid_token")) {
+                                    Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.auth_sync_fail));
+                                } else if (output.equals("nothing")) {
+                                    Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.fetch_sync_fail));
+                                } else if (!output.equals("dberror")) {
+                                    for (int i = 0; i < updatedStudies.size(); i++) {
+                                        Log.i("STUDIES MODIFIED", "notification data changed for " + updatedStudies.size() + " studies");
+                                        cancelStudy(oldStudies.get(i), false, false);
+                                        setUpNewStudyAlarms(updatedStudies.get(i));
                                     }
-                                    cancelStudy(s);
+                                    for (Study s : cancelledStudies) {
+                                        Log.i("STUDIES CANCELLED", "removed " + cancelledStudies.size() + " studies");
+                                        for (Study ks : allStudies) {
+                                            if (ks.getId() == s.getId()) {
+                                                allStudies.remove(ks);
+                                                break;
+                                            }
+                                        }
+                                        cancelStudy(s, true, true);
+                                    }
+                                    Log.i("Study sync:", context.getApplicationContext().getString(R.string.update_success));
+                                } else {
+                                    Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.fetch_sync_fail));
                                 }
-                                Log.i("Study sync:", context.getApplicationContext().getString(R.string.update_success));
-                            } else {
-                                Log.i("FINISHED SYNC:", context.getApplicationContext().getString(R.string.fetch_sync_fail));
                             }
-                        }
-                    });
-                    ExecutorSupplier.getInstance().forBackgroundTasks().execute(syncStudyDataTask);
+                        });
+                        ExecutorSupplier.getInstance().forBackgroundTasks().execute(syncStudyDataTask);
 
-                }
-            });
+                    }
+                });
 
-            ExecutorSupplier.getInstance().forBackgroundTasks().execute(syncResultDataTask);
-
+                ExecutorSupplier.getInstance().forBackgroundTasks().execute(syncResultDataTask);
+            }
 
         } else {
-            Log.i("NETWORK STATE CHANGED", "NO CONNECTION AVAILABLE");
+            Log.i("NETWORK STATE CHANGED", "DON'T SYNC");
         }
     }
 
-    public void cancelStudy(final Study study) {
+    public void cancelStudy(final Study study, final boolean cancelEvents, final boolean removeStudy) {
         mHandler.post(new Runnable() {
             @Override
             public void run() {
@@ -99,24 +115,28 @@ public class NetworkChangeReceiver extends BroadcastReceiver {
                 PendingIntent pendingIntent = PendingIntent.getBroadcast(mContext.getApplicationContext(), (int) study.getId(), new Intent(mContext.getApplicationContext(), ResponseReceiver.class), 0);
                 AlarmManager am = (AlarmManager) mContext.getApplicationContext().getSystemService(Context.ALARM_SERVICE);
                 am.cancel(pendingIntent);
-                for (Event event : EventDialogFragment.activeEvents) {
-                    if (event.getStudyId() == study.getId()) {
-                        Intent stopIntent = new Intent(mContext.getApplicationContext(), StopReceiver.class);
-                        stopIntent.putExtra("start", event.getStartTimeCalendar());
-                        stopIntent.putExtra("notificationId", EventDialogFragment.uniqueValueMap.get((int) event.getId()));
-                        stopIntent.putExtra("studyId", event.getStudyId());
-                        stopIntent.putExtra("controlNotificationId", EventDialogFragment.uniqueControlValueMap.get((int) event.getId()));
-                        stopIntent.putExtra("eventId", event.getId());
-                        mContext.getApplicationContext().sendBroadcast(stopIntent);
+                if(cancelEvents) {
+                    for (Event event : EventDialogFragment.activeEvents) {
+                        if (event.getStudyId() == study.getId()) {
+                            Intent stopIntent = new Intent(mContext.getApplicationContext(), StopReceiver.class);
+                            stopIntent.putExtra("start", event.getStartTimeCalendar());
+                            stopIntent.putExtra("notificationId", EventDialogFragment.uniqueValueMap.get((int) event.getId()));
+                            stopIntent.putExtra("studyId", event.getStudyId());
+                            stopIntent.putExtra("controlNotificationId", EventDialogFragment.uniqueControlValueMap.get((int) event.getId()));
+                            stopIntent.putExtra("eventId", event.getId());
+                            mContext.getApplicationContext().sendBroadcast(stopIntent);
+                        }
                     }
                 }
-                DBHandler.getInstance(mContext.getApplicationContext()).deleteStudyEntry(study.getId());
+                if(removeStudy) {
+                    DBHandler.getInstance(mContext.getApplicationContext()).deleteStudyEntry(study.getId());
+                }
             }
         });
     }
 
 
-    private void setUpNewStudy(final Study s) {
+    private void setUpNewStudyAlarms(final Study s) {
         mHandler.post(new Runnable() {
             @Override
             public void run() {
